Fetch correct redirect target
The API provides no good way to fetch the redirect target of a given
page without actually redirecting that page, so we find the redirect
target via a 2nd API request. The old method of using the links table
only worked if there were no other links on the redirect page, otherwise
the links were returned in alphabetical order rather than in the order
they appeared on the page.

Fixes #11