Split updateRemotePage.php
The script is now designed to be run over multiple invocations to
alleviate pressure discovered while running it in one go (memory usage
got quite a bit inflated as there are some "leaks" in MediaWiki).
Additionally the bulk of the SQL is no longer run through the MW DB
layer and instead is written to a well-known path for the user to run
separately. This should provide significantly increased performance
compared to MW's DB layer executing the file line-by-line.