More efficient method of setting mirror indicator
Previously handling only in BeforePageDisplay caused icon to not appear
in various situations, which was a bit odd because the mirror data
should have been cached. This sets the URL where we have easy access to
it, and allows us to bypass repeating a bunch of checks and potentially
database lookups later.