Add progress save feature
This allows us to resume where we left off after script failure, which
seems to happen with some frequency unfortunately. I will likely need to
add in some retry logic but I'd like to figure out if the underlying
error cause is on my end or the WME API end first.